refactor: 外移 Gateway 内联测试模块
将 gateway/stream、gateway/dispatcher、api 三个文件的尾部 #[cfg(test)]
mod tests 整体外移到同级 tests.rs，业务代码与测试断言语义保持不变。
文件由 file.rs 调整为 file/mod.rs + file/tests.rs 目录布局。

refs #109
